+++ app/package.json
@@ -38,6 +38,7 @@
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "~0.63.4",
+    "react-native-exit-app": "^1.1.0",
     "react-native-gesture-handler": "~1.8.0",
     "react-native-permissions": "^3.0.1",
     "react-native-reanimated": "~1.13.0",
@@ -46,7 +47,8 @@
     "react-native-unimodules": "^0.12.0",
     "react-native-web": "~0.13.12",
     "react-redux": "^7.2.2",
-    "redux": "^4.0.5"
+    "redux": "^4.0.5",
+    "watchman": "^1.0.0"
   },
   "devDependencies": {
     "@babel/core": "~7.9.0",
